Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main(){
-    int x;
-    int y;
-    int z;
-    scanf("%d %d %d",&x,&y,&z);
-    int a=(x+y+z)/3;
+    float x;
+    float y;
+    float z;
+    scanf("%f %f %f",&x,&y,&z);
+    float a=(x+y+z)/3;
     printf("Average: %.2f",a);
     return 0;
 }
